+++ Localization/frFR.lua
@@ -149,6 +149,7 @@ L.Settings.ClearOnLeave = "Effacer le groupe en quittant le groupe"
 L.Settings.AutoRemoveRoles = "Retirer automatiquement ces rôles"
 L.Settings.SpotlightedHeader = "Mis en avant"
 L.Settings.UnrosteredHeader = "Non mis en avant"
+L.Settings.UnrosteredRoleFilter = "Afficher ces rôles..."
 L.Settings.AddSpacer = "Ajouter un espace"
 L.Settings.NoSlots =
 "Aucun joueur mis en avant pour l'instant. Ajoutez un membre du groupe depuis la liste voisine, ou faites-en glisser un jusqu'ici."